// src/home/CustomHomeSelect/types.ts
/*
 * ====================[ START FS LICENSE ]====================
 * © 2019 FULLSAVE.
 * Ce document ne fait pas partie du domaine public. Il est et demeure la
 * propriété exclusive de FULLSAVE.
 * Il est protégé par le Code de la Propriété intellectuelle et le droit
 * d’auteur.
 * Toute reproduction est interdite. Toute représentation est interdite.
 * Tous droits réservés. Droits non cessibles.
 * =====================[ END FS LICENSE ]=====================
 */

export type TCookie = {
  label: string;
  value: string;
};

export type TConfig = {
  label: string;
  value: string;
  tools: string[];
  toolsProd: string[];
};